remote unused code & add socks5 logic
