--- m_server/socks_conn.go
@@ -1,391 +1,0 @@
-package m_server
-
-import (
-	"encoding/base64"
-	"errors"
-	"io"
-	"net"
-	"os"
-	"strconv"
-	"sync"
-	"time"
-)
-
-import (
-	"github.com/baidu/go-lib/gotrack"
-	"github.com/baidu/go-lib/log"
-	"github.com/zyong/miniproxygo/m_core"
-)
-
-
-// This should be >= 512 bytes for DetectContentType,
-// but otherwise it's somewhat arbitrary.
-const bufferBeforeChunkingSize = 512
-
-// MaxAddrLen is the maximum size of SOCKS address in bytes.
-const MaxAddrLen = 1 + 1 + 255 + 2
-
-// Actions to do with current connection.
-const (
-	// Reuse the connection (default aciton)
-	keepAlive = iota
-	// Close the connection after send response to client.
-	closeAfterReply
-	// Close the connection directly, do not send any data,
-	// it usually means some attacks may happened.
-	closeDirectly
-)
-
-
-// UDPEnabled is the toggle for UDP support
-var UDPEnabled = false
-
-// SOCKS request commands as defined in RFC 1928 section 4.
-const (
-	CmdConnect      = 1
-	CmdBind         = 2
-	CmdUDPAssociate = 3
-)
-
-// SOCKS address types as defined in RFC 1928 section 5.
-const (
-	AtypIPv4       = 1
-	AtypDomainName = 3
-	AtypIPv6       = 4
-)
-
-// Addr represents a SOCKS address as defined in RFC 1928 section 5.
-type Addr []byte
-
-
-// String serializes SOCKS address a to string form.
-func (a Addr) String() string {
-	var host, port string
-
-	switch a[0] { // address type
-	case AtypDomainName:
-		host = string(a[2 : 2+int(a[1])])
-		port = strconv.Itoa((int(a[2+int(a[1])]) << 8) | int(a[2+int(a[1])+1]))
-	case AtypIPv4:
-		host = net.IP(a[1 : 1+net.IPv4len]).String()
-		port = strconv.Itoa((int(a[1+net.IPv4len]) << 8) | int(a[1+net.IPv4len+1]))
-	case AtypIPv6:
-		host = net.IP(a[1 : 1+net.IPv6len]).String()
-		port = strconv.Itoa((int(a[1+net.IPv6len]) << 8) | int(a[1+net.IPv6len+1]))
-	}
-
-	return net.JoinHostPort(host, port)
-}
-
-
-// Error represents a SOCKS error
-type Error byte
-
-func (err Error) Error() string {
-	return "SOCKS error: " + strconv.Itoa(int(err))
-}
-
-// SOCKS errors as defined in RFC 1928 section 6.
-const (
-	ErrGeneralFailure       = Error(1)
-	ErrConnectionNotAllowed = Error(2)
-	ErrNetworkUnreachable   = Error(3)
-	ErrHostUnreachable      = Error(4)
-	ErrConnectionRefused    = Error(5)
-	ErrTTLExpired           = Error(6)
-	ErrCommandNotSupported  = Error(7)
-	ErrAddressNotSupported  = Error(8)
-	InfoUDPAssociate        = Error(9)
-)
-
-
-var errTooLarge = errors.New("http: request too large")
-
-// A switchReader can have its Reader changed at runtime.
-// It's not safe for concurrent Reads and switches.
-type switchReader struct {
-	io.Reader
-}
-
-// A liveSwitchReader is a switchReader that's safe for concurrent
-// reads and switches, if its mutex is held.
-type liveSwitchReader struct {
-	sync.Mutex
-	r io.Reader
-}
-
-func (sr *liveSwitchReader) Read(p []byte) (n int, err error) {
-	sr.Lock()
-	r := sr.r
-	sr.Unlock()
-	return r.Read(p)
-}
-
-
-// A conn represents the server side of an HTTP/HTTPS connection.
-type conn struct {
-	// immutable:
-	remoteAddr string             // network address of remote side
-	server     *Server         // the Server on which the connection arrived
-	rwc        net.Conn           // i/o connection
-
-	// for http/https:
-	sr    liveSwitchReader      // where the LimitReader reads from; usually the rwc
-	lr    *io.LimitedReader     // io.LimitReader(sr)
-	reqSN uint32                //number of requests arrived on this connection
-
-	mu           sync.Mutex // guards the following
-	clientGone   bool       // if client has disconnected mid-request
-	closeNotifyc chan bool  // made lazily
-}
-
-func (c *conn) closeNotify() <-chan bool {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.closeNotifyc == nil {
-		c.closeNotifyc = make(chan bool, 1)
-
-		pr, pw := io.Pipe()
-
-		readSource := c.sr.r
-		c.sr.Lock()
-		c.sr.r = pr
-		c.sr.Unlock()
-		go func() {
-			_, err := io.Copy(pw, readSource)
-			if err == nil {
-				err = io.EOF
-			}
-			pw.CloseWithError(err)
-			c.noteClientGone()
-		}()
-	}
-	return c.closeNotifyc
-}
-
-func (c *conn) noteClientGone() {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	if c.closeNotifyc != nil && !c.clientGone {
-		c.closeNotifyc <- true
-	}
-	c.clientGone = true
-}
-
-func (c *conn) serve() {
-	c.server.connWaitGroup.Add(1)
-
-	defer func() {
-		if err := recover(); err != nil {
-			log.Logger.Warn("panic: conn.serve(): %v, %v\n%s",
-				c.remoteAddr,
-				err,
-				gotrack.CurrentStackTrace(0))
-		}
-	}()
-
-	defer func() {
-		c.close()
-	}()
-
-	// first request
-	tgt, err := c.HandShake(c.rwc)
-
-	rc, err := net.Dial("tcp", c.server.Config.Server.RemoteServer)
-	if err != nil {
-		log.Logger.Warn("failed to connect to server %v: %v", c.server.Config.Server.RemoteServer, err)
-		return
-	}
-
-	defer rc.Close()
-
-	k, err := base64.URLEncoding.DecodeString(c.server.Config.Server.Key)
-	if err != nil {
-		log.Logger.Warn("failed to decode server key")
-		return
-	}
-	ciph, err := m_core.PickCipher(c.server.Config.Server.Cipher, k, c.server.Config.Server.Password)
-	rc = ciph.StreamConn(rc)
-
-	if _, err = rc.Write(tgt); err != nil {
-		log.Logger.Warn("failed to send target address: %v", err)
-		return
-	}
-
-	log.Logger.Info("proxy %s <-> %s <-> %s", c.rwc.RemoteAddr(), c.server.Config.Server.RemoteServer, tgt)
-
-	if err = relay(rc, c.rwc); err != nil {
-		log.Logger.Warn("relay error: %v", err)
-	}
-
-}
-
-
-// relay copies between left and right bidirectionally
-func relay(left, right net.Conn) error {
-	var err, err1 error
-	var wg sync.WaitGroup
-	var wait = 5 * time.Second
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		_, err1 = io.Copy(right, left)
-		right.SetReadDeadline(time.Now().Add(wait)) // unblock read on right
-	}()
-	_, err = io.Copy(left, right)
-	left.SetReadDeadline(time.Now().Add(wait)) // unblock read on left
-	wg.Wait()
-	if err1 != nil && !errors.Is(err1, os.ErrDeadlineExceeded) { // requires Go 1.15+
-		return err1
-	}
-	if err != nil && !errors.Is(err, os.ErrDeadlineExceeded) {
-		return err
-	}
-	return nil
-}
-
-
-// ParseAddr parses the address in string s. Returns nil if failed.
-func ParseAddr(s string) Addr {
-	var addr Addr
-	host, port, err := net.SplitHostPort(s)
-	if err != nil {
-		return nil
-	}
-	if ip := net.ParseIP(host); ip != nil {
-		if ip4 := ip.To4(); ip4 != nil {
-			addr = make([]byte, 1+net.IPv4len+2)
-			addr[0] = AtypIPv4
-			copy(addr[1:], ip4)
-		} else {
-			addr = make([]byte, 1+net.IPv6len+2)
-			addr[0] = AtypIPv6
-			copy(addr[1:], ip)
-		}
-	} else {
-		if len(host) > 255 {
-			return nil
-		}
-		addr = make([]byte, 1+1+len(host)+2)
-		addr[0] = AtypDomainName
-		addr[1] = byte(len(host))
-		copy(addr[2:], host)
-	}
-
-	portnum, err := strconv.ParseUint(port, 10, 16)
-	if err != nil {
-		return nil
-	}
-
-	addr[len(addr)-2], addr[len(addr)-1] = byte(portnum>>8), byte(portnum)
-
-	return addr
-}
-
-func (c *conn) HandShake(rw io.ReadWriter) ([]byte, error) {
-	// Read RFC 1928 for request and reply structure and sizes.
-	buf := make([]byte, MaxAddrLen)
-	// read VER, NMETHODS, METHODS
-	if _, err := io.ReadFull(rw, buf[:2]); err != nil {
-		return nil, err
-	}
-	nmethods := buf[1]
-	if _, err := io.ReadFull(rw, buf[:nmethods]); err != nil {
-		return nil, err
-	}
-	// write VER METHOD
-	if _, err := rw.Write([]byte{5, 0}); err != nil {
-		return nil, err
-	}
-	// read VER CMD RSV ATYP DST.ADDR DST.PORT
-	if _, err := io.ReadFull(rw, buf[:3]); err != nil {
-		return nil, err
-	}
-	cmd := buf[1]
-	addr, err := readAddr(rw, buf)
-	if err != nil {
-		return nil, err
-	}
-	switch cmd {
-	case CmdConnect:
-		_, err = rw.Write([]byte{5, 0, 0, 1, 0, 0, 0, 0, 0, 0}) // SOCKS v5, reply succeeded
-	case CmdUDPAssociate:
-		if !UDPEnabled {
-			return nil, ErrCommandNotSupported
-		}
-		listenAddr := ParseAddr(rw.(net.Conn).LocalAddr().String())
-		_, err = rw.Write(append([]byte{5, 0, 0}, listenAddr...)) // SOCKS v5, reply succeeded
-		if err != nil {
-			return nil, ErrCommandNotSupported
-		}
-		err = InfoUDPAssociate
-	default:
-		return nil, ErrCommandNotSupported
-	}
-
-	return addr, err // skip VER, CMD, RSV fields
-}
-
-
-func readAddr(r io.Reader, b []byte) ([]byte, error) {
-	if len(b) < MaxAddrLen {
-		return nil, io.ErrShortBuffer
-	}
-	_, err := io.ReadFull(r, b[:1]) // read 1st byte for address type
-	if err != nil {
-		return nil, err
-	}
-
-	switch b[0] {
-	case AtypDomainName:
-		_, err = io.ReadFull(r, b[1:2]) // read 2nd byte for domain length
-		if err != nil {
-			return nil, err
-		}
-		_, err = io.ReadFull(r, b[2:2+int(b[1])+2])
-		return b[:1+1+int(b[1])+2], err
-	case AtypIPv4:
-		_, err = io.ReadFull(r, b[1:1+net.IPv4len+2])
-		return b[:1+net.IPv4len+2], err
-	case AtypIPv6:
-		_, err = io.ReadFull(r, b[1:1+net.IPv6len+2])
-		return b[:1+net.IPv6len+2], err
-	}
-
-	return nil, ErrAddressNotSupported
-}
-
-// ReadAddr reads just enough bytes from r to get a valid Addr.
-func ReadAddr(r io.Reader) ([]byte, error) {
-	return readAddr(r, make([]byte, MaxAddrLen))
-}
-
-
-// noLimit is an effective infinite upper bound for io.LimitedReader
-const noLimit int64 = (1 << 63) - 1
-
-// Create new connection from rwc.
-func newConn(rwc net.Conn, srv *Server) (c *conn, err error) {
-	c = new(conn)
-	c.remoteAddr = rwc.RemoteAddr().String()
-	c.server = srv
-	c.rwc = rwc
-
-	return c, nil
-}
-
-// Close the connection.
-func (c *conn) close() {
-	c.rwc.Close()
-}
-
-
-
-// rstAvoidanceDelay is the amount of time we sleep after closing the
-// write side of a TCP connection before closing the entire socket.
-// By sleeping, we increase the chances that the client sees our FIN
-// and processes its final data before they process the subsequent RST
-// from closing a connection with known unread data.
-// This RST seems to occur mostly on BSD systems. (And Windows?)
-// This timeout is somewhat arbitrary (~latency around the planet).
-const rstAvoidanceDelay = 500 * time.Millisecond
